Update index.md
Remove ambigous symbols, and adjust colors for thematic consistency

The Apple-Shortcut and Unix indicators in the tools-tab were ambigous and led to confusion. They didn't accurately portray the target - Shortcuts the Apple-app, Unix as in relation to Linux (similarities).

Apple-Shortcuts may be added at a later date if a proper indicator is acquired.